+++ crates/stc_ts_file_analyzer/src/analyzer/mod.rs
@@ -65,7 +65,6 @@ mod hoisting;
 mod import;
 mod pat;
 mod props;
-mod relation;
 mod scope;
 mod stmt;
 #[cfg(test)]
